requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0
tensorboard>=2.13.0

# NLP and Medical NLP
transformers>=4.30.0
datasets>=2.14.0
spacy>=3.6.0
scispacy>=0.5.3
medspacy>=1.0.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Additional NLP utilities
seqeval>=1.2.2
conllu>=4.5.0
nltk>=3.8.0

# Data processing
omegaconf>=2.3.0
hydra-core>=1.3.0

# Visualization and demos
streamlit>=1.25.0
gradio>=3.40.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
ruff>=0.0.280
pre-commit>=3.3.0

# Optional: Advanced NLP models
# torch-audio>=2.0.0  # For speech processing
# wandb>=0.15.0      # For experiment tracking
# mlflow>=2.5.0      # Alternative experiment tracking
